Refactor: Resolve conversasion in code review
