+++ src/app/lib/hooks/previous.jsx
@@ -1,0 +1,10 @@
+import { useRef, useEffect } from 'react';
+
+function usePrevious(value) {
+    const ref = useRef();
+    useEffect(() => {
+        ref.current = value;
+    });
+    return ref.current;
+}
+export default usePrevious;
