+++ src/app/ui/pages/Workspace.jsx
@@ -2,7 +2,7 @@ import _ from 'lodash';
 import classNames from 'classnames';
 import React, { useEffect, useRef, useState } from 'react';
 import includes from 'lodash/includes';
-import { useHistory, withRouter } from 'react-router-dom';
+import { useHistory } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
 import PropTypes from 'prop-types';
 import { Button } from '../components/Buttons';
@@ -274,4 +274,4 @@ Workspace.propTypes = {
     className: PropTypes.string
 };
 
-export default (withRouter(Workspace));
+export default Workspace;
